# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetProvider": {
        "description": "Delivery provider for a derived, renderable asset.\n\nCesium ion is the current delivery provider; it is deliberately *not* the\ncanonical data model (see docs/ARCHITECTURE.md).",
        "enum": [
          "cesium-ion",
          "3d-tiles-url"
        ],
        "title": "AssetProvider",
        "type": "string"
      },
      "AssetRead": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "title": "Attribution",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "crs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrsMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "defaultVisible": {
            "default": false,
            "title": "Defaultvisible",
            "type": "boolean"
          },
          "footprint": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "MultiPolygon": "#/components/schemas/MultiPolygon",
                    "Polygon": "#/components/schemas/Polygon"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Polygon"
                  },
                  {
                    "$ref": "#/components/schemas/MultiPolygon"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Footprint"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "observedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider": {
            "$ref": "#/components/schemas/AssetProvider"
          },
          "renderConfig": {
            "$ref": "#/components/schemas/RenderConfig"
          },
          "representation": {
            "$ref": "#/components/schemas/Representation"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "siteId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siteid"
          },
          "source": {
            "discriminator": {
              "mapping": {
                "3d-tiles-url": "#/components/schemas/TilesUrlSource",
                "cesium-ion": "#/components/schemas/CesiumIonSource"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CesiumIonSource"
              },
              {
                "$ref": "#/components/schemas/TilesUrlSource"
              }
            ],
            "title": "Source"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "validFrom": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validfrom"
          },
          "validTo": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validto"
          }
        },
        "required": [
          "name",
          "representation",
          "source",
          "id",
          "siteId",
          "provider",
          "createdAt",
          "updatedAt"
        ],
        "title": "AssetRead",
        "type": "object"
      },
      "AssetUpdate": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Attribution"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "defaultVisible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultvisible"
          },
          "footprint": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "MultiPolygon": "#/components/schemas/MultiPolygon",
                    "Polygon": "#/components/schemas/Polygon"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Polygon"
                  },
                  {
                    "$ref": "#/components/schemas/MultiPolygon"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Footprint"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "observedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "renderConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenderConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "validFrom": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validfrom"
          },
          "validTo": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validto"
          }
        },
        "title": "AssetUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IonAssetMetadata": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          },
          "dateAdded": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateadded"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "percentComplete": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentcomplete",
            "type": "integer"
          },
          "status": {
            "enum": [
              "AWAITING_FILES",
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETE",
              "ERROR",
              "DATA_ERROR"
            ],
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "status",
          "percentComplete"
        ],
        "title": "IonAssetMetadata",
        "type": "object"
      },
...
      "LayerCreate": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "title": "Attribution",
            "type": "array"
          },
          "category": {
            "$ref": "#/components/schemas/LayerCategory"
          },
          "coverage": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "defaultVisible": {
            "default": false,
            "title": "Defaultvisible",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "legend": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegendMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "observedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "render": {
            "$ref": "#/components/schemas/RenderMetadata"
          },
          "resolution": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "slug": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "source": {
            "discriminator": {
              "mapping": {
                "3d-tiles-url": "#/components/schemas/TilesUrlLayerSource",
                "arcgis-mapserver": "#/components/schemas/ArcGisMapServerSource",
                "cesium-ion-3d-tiles": "#/components/schemas/CesiumIon3DTilesSource",
                "cesium-ion-imagery": "#/components/schemas/CesiumIonImagerySource",
                "cesium-ion-terrain": "#/components/schemas/CesiumIonTerrainSource",
                "czml": "#/components/schemas/CzmlSource",
                "geojson": "#/components/schemas/GeoJsonSource",
                "google-photorealistic": "#/components/schemas/GooglePhotorealisticSource",
                "mvt": "#/components/schemas/MvtSource",
                "stac": "#/components/schemas/StacSource",
                "wms": "#/components/schemas/WmsSource",
                "wmts": "#/components/schemas/WmtsSource",
                "xyz": "#/components/schemas/XyzSource"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CesiumIonTerrainSource"
              },
              {
                "$ref": "#/components/schemas/CesiumIonImagerySource"
              },
              {
                "$ref": "#/components/schemas/CesiumIon3DTilesSource"
              },
              {
                "$ref": "#/components/schemas/GooglePhotorealisticSource"
              },
              {
                "$ref": "#/components/schemas/TilesUrlLayerSource"
              },
              {
                "$ref": "#/components/schemas/XyzSource"
              },
              {
                "$ref": "#/components/schemas/WmtsSource"
              },
              {
                "$ref": "#/components/schemas/WmsSource"
              },
              {
                "$ref": "#/components/schemas/ArcGisMapServerSource"
              },
              {
                "$ref": "#/components/schemas/GeoJsonSource"
              },
              {
                "$ref": "#/components/schemas/CzmlSource"
              },
              {
                "$ref": "#/components/schemas/MvtSource"
              },
              {
                "$ref": "#/components/schemas/StacSource"
              }
            ],
            "title": "Source"
          },
          "spatialExtent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "temporalExtent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemporalExtent"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "category",
          "source"
        ],
        "title": "LayerCreate",
        "type": "object"
      },
      "LayerRead": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "title": "Attribution",
            "type": "array"
          },
          "builtin": {
            "title": "Builtin",
            "type": "boolean"
          },
          "category": {
            "$ref": "#/components/schemas/LayerCategory"
          },
          "coverage": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "defaultVisible": {
            "default": false,
            "title": "Defaultvisible",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "legend": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegendMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "observedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "render": {
            "$ref": "#/components/schemas/RenderMetadata"
          },
          "resolution": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source": {
            "discriminator": {
              "mapping": {
                "3d-tiles-url": "#/components/schemas/TilesUrlLayerSource",
                "arcgis-mapserver": "#/components/schemas/ArcGisMapServerSource",
                "cesium-ion-3d-tiles": "#/components/schemas/CesiumIon3DTilesSource",
                "cesium-ion-imagery": "#/components/schemas/CesiumIonImagerySource",
                "cesium-ion-terrain": "#/components/schemas/CesiumIonTerrainSource",
                "czml": "#/components/schemas/CzmlSource",
                "geojson": "#/components/schemas/GeoJsonSource",
                "google-photorealistic": "#/components/schemas/GooglePhotorealisticSource",
                "mvt": "#/components/schemas/MvtSource",
                "stac": "#/components/schemas/StacSource",
                "wms": "#/components/schemas/WmsSource",
                "wmts": "#/components/schemas/WmtsSource",
                "xyz": "#/components/schemas/XyzSource"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CesiumIonTerrainSource"
              },
              {
                "$ref": "#/components/schemas/CesiumIonImagerySource"
              },
              {
                "$ref": "#/components/schemas/CesiumIon3DTilesSource"
              },
              {
                "$ref": "#/components/schemas/GooglePhotorealisticSource"
              },
              {
                "$ref": "#/components/schemas/TilesUrlLayerSource"
              },
              {
                "$ref": "#/components/schemas/XyzSource"
              },
              {
                "$ref": "#/components/schemas/WmtsSource"
              },
              {
                "$ref": "#/components/schemas/WmsSource"
              },
              {
                "$ref": "#/components/schemas/ArcGisMapServerSource"
              },
              {
                "$ref": "#/components/schemas/GeoJsonSource"
              },
              {
                "$ref": "#/components/schemas/CzmlSource"
              },
              {
                "$ref": "#/components/schemas/MvtSource"
              },
              {
                "$ref": "#/components/schemas/StacSource"
              }
            ],
            "title": "Source"
          },
          "sourceType": {
            "$ref": "#/components/schemas/LayerSourceType"
          },
          "spatialExtent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "temporalExtent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemporalExtent"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "name",
          "category",
          "source",
          "id",
          "slug",
          "sourceType",
          "builtin",
          "createdAt",
          "updatedAt"
        ],
        "title": "LayerRead",
        "type": "object"
      },
...
      "Representation": {
        "enum": [
          "gaussian-splat",
          "mesh",
          "point-cloud",
          "terrain",
          "imagery"
        ],
        "title": "Representation",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "\nCatalog service for the geospatial digital twin.\n\n* **Sites** \u2014 physical places with reality captures, footprints and camera bookmarks.\n* **Assets** \u2014 derived, renderable delivery assets (3D Tiles) with provenance.\n* **Layers** \u2014 composable world layers from open data or the user's own sources.\n\nAll geometry is GeoJSON (WGS 84, RFC 7946). All timestamps are ISO 8601.\n",
    "title": "Digital Twin API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assets/{asset_id}": {
      "delete": {
        "operationId": "delete_asset_api_v1_assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an asset",
        "tags": [
          "assets"
        ]
      },
      "get": {
        "operationId": "get_asset_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get an asset",
        "tags": [
          "assets"
        ]
      },
      "patch": {
        "operationId": "update_asset_api_v1_assets__asset_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update an asset",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/v1/ion/assets/{asset_id}": {
      "get": {
        "description": "Narrow, documented read of `GET https://api.cesium.com/v1/assets/{id}` using the server-side ion token. Used to monitor reconstruction jobs. This is not a general proxy.",
        "operationId": "ion_asset_api_v1_ion_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IonAssetMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read ion asset metadata / tiling status",
        "tags": [
          "cesium-ion"
        ]
      }
    },
...
    "/api/v1/layers": {
      "get": {
        "operationId": "list_layers_api_v1_layers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LayerRead"
                  },
                  "title": "Response List Layers Api V1 Layers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List catalog layers",
        "tags": [
          "layers"
        ]
      },
      "post": {
        "operationId": "create_layer_api_v1_layers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a layer",
        "tags": [
          "layers"
        ]
      }
    },
    "/api/v1/layers/{layer_id}": {
      "delete": {
        "operationId": "delete_layer_api_v1_layers__layer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "layer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Layer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a layer",
        "tags": [
          "layers"
        ]
      },
      "get": {
        "operationId": "get_layer_api_v1_layers__layer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "layer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Layer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a layer",
        "tags": [
          "layers"
        ]
      },
      "patch": {
        "operationId": "update_layer_api_v1_layers__layer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "layer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Layer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a layer",
        "tags": [
          "layers"
        ]
      }
    },
...
  }
}